requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-dateutil==2.8.2
SQLAlchemy==2.0.23
alembic==1.12.1
flask_sqlalchemy==3.1.1
